fix(bench): drop useless static_cast on state.iterations()
Google Benchmark's IterationCount typedef alias is std::int64_t on libstdc++; wrapping state.iterations() in static_cast<std::int64_t> trips -Wuseless-cast under GCC. Keep the cast on the std::size_t multiplier (still needed for -Wsign-conversion) and drop it on the iteration count.

+++ bench/bench_bitmap.cpp
@@ -36,7 +36,7 @@ void bench_set(benchmark::State& state) {
         benchmark::DoNotOptimize(bm);
         bm.clear_all();
     }
-    state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+    state.SetItemsProcessed(state.iterations() *
                             static_cast<std::int64_t>(bits.size()));
 }
 
@@ -54,7 +54,7 @@ void bench_set_clear_pair(benchmark::State& state) {
             bm.clear(b);
         benchmark::DoNotOptimize(bm);
     }
-    state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+    state.SetItemsProcessed(state.iterations() *
                             static_cast<std::int64_t>(bits.size()));
 }
